Move Discovery Format support to a top-level subpackage.
