+++ discovery/README.md
@@ -1,0 +1,5 @@
+# Discovery
+
+This directory contains utility code for reading API descriptions in
+Google's [Discovery Format](https://developers.google.com/discovery/v1/reference/apis)
+and converting them to OpenAPI.
